Emenda de novas assinaturas funcionando.
OBS: It seems that the OpenSSL team has not thought much about this use case, as calling CMS_final fails beacuse we don't have
the private key of the previous signature. Not calling it, however, results in the message digest not being calculated for the
new signaure. The "solution" (read hack) is not use CMS_REUSE_DIGEST and NOT calling CMS_final.

I tried to look into the OpenSSL source to understand how to calculate the message digest without calling CMS_final but I just
could not find how.